0123 GraduationProject_Web/flask_api/requirements.txt
flask==2.3.3
pillow==10.0.0
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
cors==1.0.1
flask-cors==4.0.0 